refactor: colocate dynamic ssr:false boundary on ScheduleManagementContent
